BTextView: Fix Tracker edit name
Don't resize text view in FrameResized() if resizable, this is done
in _AutoResize() instead. Scroll to 0 after resizing in _AutoResize().

Set TextRect width to max line width when word-wrap is off.
Make _ValidateTextRect() method and call it to ensure fTextRect
does not reach 0 width. This change means that the text rect width is
always the width of the text inside matching BeOS R5 behavior. It is
necessary to allow the Edit Name text view to shrink as text is
deleted.

TODO highlighting only goes to edge of text, it should go to the edge
of text view when wrap is off (e.g. in StyledEdit).

Prevent scrolling out of bounds horizontally with the keyboard.
The algorithm to keep cursor in text view bounds was taken from
_PerformAutoScrolling() (mouse scrolling method) and was updated
to account for the left margin which is not always 0.

Make left and top margin not always 0 when scrolling in
_PerformAutoScrolling() as well, this fixes the same bug when
scrolling the text with the mouse.

Remove extraSpace variable, no longer needed. This means that we
always scroll according to the text width without any extra space.

The above makes text views behave properly when overflow occurs i.e.
when you type text off the end of the text view. Instead of scrolling
an entire page of text over, it nudges the text over slightly as
you'd expect it to.

Take top and bottom margins into account when scrolling with keyboard
in ScrollToOffset(). This allows you to scroll past the margins all the
way to the top and bottom using the up and down arrow keys (1 more px.)

Fix drawing issues when typing off the edge:

When word-wrap is turned on for a multi-line text view and you insert,
backspace or delete the position of text rect changes as you type off
the end of the BTextView area. Call FrameResized() in ScrollToOffset()
to reposition the text rect and redraw. This can also happen as a result
of the word-wrap setting changing too so call FrameResized() in that
case too.

Use Window()->DisableUpdates() and Window()->EnableUpdates() and
remove a call to Window()->UpdateIfNeeded() to prevent the text
from flashing when it draw then the text rect position is moved
and it draws again. This coalesces multiple consecutive draw calls.

Replace xDiff and yDiff float variables with a single scrollTo BPoint
in ScrollToOffset(). This was also copied from _PerformAutoScrolling().
(no functional change intended here)

Remove floorf() calls, we know that we are dealing with integers.
(no functional change intended here)

Fixes #16642, #16476

Change-Id: I2d32d6039944d2dc3218ce4de71f2966cc98c866